Issue #91: 8051 ajmp / acall instructions need to ignore upper bits and check address based on address + 2.

+++ disasm/8051.h
@@ -5,7 +5,7 @@
  *     Web: http://www.mikekohn.net/
  * License: GPLv3
  *
- * Copyright 2010-2019 by Michael Kohn
+ * Copyright 2010-2021 by Michael Kohn
  *
  */
 
@@ -14,10 +14,25 @@
 
 #include "common/assembler.h"
 
-int get_cycle_count_8051(unsigned short int opcode);
-int disasm_8051(struct _memory *memory, uint32_t address, char *instruction, int *cycles_min, int *cycles_max);
-void list_output_8051(struct _asm_context *asm_context, uint32_t start, uint32_t end);
-void disasm_range_8051(struct _memory *memory, uint32_t flags, uint32_t start, uint32_t end);
+int get_cycle_count_8051(uint32_t opcode);
+
+int disasm_8051(
+  struct _memory *memory,
+  uint32_t address,
+  char *instruction,
+  int *cycles_min,
+  int *cycles_max);
+
+void list_output_8051(
+  struct _asm_context *asm_context,
+  uint32_t start,
+  uint32_t end);
+
+void disasm_range_8051(
+  struct _memory *memory,
+  uint32_t flags,
+  uint32_t start,
+  uint32_t end);
 
 #endif
